Restructure libs with helper functions
