+++ lib/forecast_utils.js
@@ -4,10 +4,21 @@
 'use strict'
 
 
+const csv = require('dwd-csv-helper')
 const moment = require('moment')
 const path = require('path')
 const _ = require('lodash')
+const gu = require('../lib/general_utils.js')
+const bunyan = require('bunyan')
+const log = bunyan.createLogger({
+  name: 'dwd-csv-handler',
+  serializers: bunyan.stdSerializers
+})
 
+
+const secInHour = 3600
+const milliseconds = 1000
+const hoursInDay = 24
 
 const cosmoD2AvailableFrom = moment.utc('2018051509', 'YYYYMMDDHH')
 
@@ -59,5 +70,93 @@ function renderMeasuredValuesAsJSON (vois, stationId, modelRun, model, startTime
   }
 
 
+  function deriveCsvFilePath (csvBasePath, type, timestamp, stationId, modelRun) {
+    // No test for this function
+    let formatString
+    let contentType
+    let extension
+    let subdir
+  
+    if (type === 'REPORT') {
+      formatString = 'YYYYMMDD'
+      contentType = 'BEOB'
+      extension = '.csv'
+      subdir = 'poi'
+    } else if (type === 'MOSMIX_KMZ') {
+      formatString = 'YYYYMMDDHH'
+      contentType = 'MOSMIX'
+      extension = '.kmz'
+      subdir = 'mos'
+    } else {
+      formatString = 'YYYYMMDDHH'
+      contentType = 'MOSMIX'
+      extension = '.csv'
+      subdir = 'poi'
+    }
+  
+    const dayDateTimeString = moment.utc(timestamp).format(formatString).slice(0, -2) + modelRun
+    const fileName = stationId + '-' + contentType + extension
+  
+    return path.join(csvBasePath, subdir, dayDateTimeString, fileName)
+  }
+  
+  async function getMosmixDataDirectories(mosmixBasePath, startTimestamp, endTimestamp, stationId, modelRun) {
+    // No test for this function
+    const result = []
+    let timestamp = startTimestamp - (secInHour * milliseconds)
+    while (timestamp <= endTimestamp) {
+      // Timestamp is increased by 24 hours in iteration
+      // Model-run is explicitly += to the dayDateTimeString
+      const filePath = deriveCsvFilePath(mosmixBasePath, 'MOSMIX_KMZ', timestamp, stationId, modelRun)
+      result.push(filePath)
+      timestamp += hoursInDay * secInHour * milliseconds
+    }
+    return result
+  }
+  
+  async function readMosmixTimeseriesData(mosmixBasePath, startTimestamp, endTimestamp, stationId, modelRun) {
+    // No test for this function
+    const resultArray = []
+    const mosmixFiles = await getMosmixDataDirectories(mosmixBasePath, startTimestamp, endTimestamp, stationId, modelRun)
+    for (let i = 0; i < mosmixFiles.length; i += 1) {
+      const filePath = mosmixFiles[i]
+      try {
+        const fileContent = await csv.extractKmlFile(filePath)
+        const result = await csv.parseKmlFile(fileContent)
+        resultArray.push(result)
+      } catch (error) {
+          log.warn(error, `failed to extract the .kml-file from ${filePath}`)
+      }
+    }
+    return resultArray
+  }
+
+  function renderMeasuredValuesAsCSV (voiConfigs, timeseriesDataArray) {
+    return gu.renderMeasuredValuesAsCSV(voiConfigs, timeseriesDataArray)
+  }
+
+  function dropTimeseriesDataNotOfInteresWithParameter (voiConfigs, timeseriesDataCollection, parameter) {
+    const result = []
+    const keys = _.map(voiConfigs, function (o) {
+      return o[parameter].key
+    })
+    _.forEach(timeseriesDataCollection, timeseriesDataItem => {
+      result.push(_.pick(timeseriesDataItem, keys))
+    })
+    return result
+  }
+
+  function renderMeasuredValuesAsCSV (voiConfigs, timeseriesDataArray) {
+    return gu.renderMeasuredValuesAsCSV(voiConfigs, timeseriesDataArray)
+  }
+
+  function checkIfValidModelRunMosmix(allowed_mosmix_mr, modelRun) {
+    return allowed_mosmix_mr.includes(modelRun)
+  }
 exports.getGribDirectory = getGribDirectory
 exports.renderMeasuredValuesAsJSON = renderMeasuredValuesAsJSON
+exports.readMosmixTimeseriesData = readMosmixTimeseriesData
+exports.renderMeasuredValuesAsCSV = renderMeasuredValuesAsCSV
+exports.dropTimeseriesDataNotOfInteresWithParameter = dropTimeseriesDataNotOfInteresWithParameter
+exports.renderMeasuredValuesAsCSV = renderMeasuredValuesAsCSV
+exports.checkIfValidModelRunMosmix = checkIfValidModelRunMosmix
